// apps/bookingservice/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const BookingStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  PAYMENT_FAILED: 'PAYMENT_FAILED',
  CONFIRMED: 'CONFIRMED',
  IN_PROGRESS: 'IN_PROGRESS',
  ARRIVAL_CONFIRMED: 'ARRIVAL_CONFIRMED',
  PENDING_COMPLETION_VERIFICATION: 'PENDING_COMPLETION_VERIFICATION',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const CancelledBy = {
  USER: 'USER',
  PROVIDER: 'PROVIDER',
  SYSTEM: 'SYSTEM',
} as const;

export type CancelledBy = (typeof CancelledBy)[keyof typeof CancelledBy];

export const BookingOtpType = {
  ARRIVAL: 'ARRIVAL',
  COMPLETION: 'COMPLETION',
} as const;

export type BookingOtpType = (typeof BookingOtpType)[keyof typeof BookingOtpType];